database-agent/src/database_agent/database_agent.py
"""Main module for the Database Agent."""

import logging
from typing import Dict, List, Optional

from opentelemetry import trace

logger = logging.getLogger(__name__)
tracer = trace.get_tracer(__name__)


class DatabaseAgent:
    """Main class for database operations and management."""

    def __init__(self, config: Optional[Dict] = None):
        """Initialize the Database Agent.

        Args:
            config: Optional configuration dictionary
        """
        self.config = config or {}
        self._setup_logging()

    def _setup_logging(self):
        """Configure logging for the agent."""
        logging.basicConfig(
            level=logging.INFO,
            format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
        )

    def analyze_database(self, database_url: str) -> Dict:
        """Analyze a database and return its structure.

        Args:
            database_url: URL or connection string for the database

        Returns:
            Dictionary containing database analysis results
        """
        with tracer.start_as_current_span("DatabaseAgent.analyze_database") as span:
            span.set_attribute("database_url", database_url)
            logger.info(f"Analyzing database at {database_url}")
            # TODO: Implement database analysis logic
            return {"status": "not_implemented"}

    def generate_migration(self, changes: List[Dict]) -> str:
        """Generate database migration script based on changes.

        Args:
            changes: List of database changes to implement

        Returns:
            Migration script as a string
        """
        with tracer.start_as_current_span("DatabaseAgent.generate_migration") as span:
            span.set_attribute("changes.count", len(changes))
            logger.info(f"Generating migration for {len(changes)} changes")
            # TODO: Implement migration generation logic
            return "-- Migration script placeholder"

    def validate_changes(self, changes: List[Dict]) -> bool:
        """Validate proposed database changes.

        Args:
            changes: List of database changes to validate

        Returns:
            True if changes are valid, False otherwise
        """
        with tracer.start_as_current_span("DatabaseAgent.validate_changes") as span:
            span.set_attribute("changes.count", len(changes))
            logger.info(f"Validating {len(changes)} changes")
            # TODO: Implement validation logic
            return True
