database-agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
PyGithub==2.6.1
qdrant-client==1.14.2
sentence-transformers==4.1.0
openai==1.79.0
requests==2.32.3
pydantic==2.11.4
scikit-learn==1.6.1
numpy==2.2.5
tqdm==4.67.1
transformers==4.51.3
starlette==0.46.2
PyYAML==6.0.2
huggingface-hub==0.31.2
tokenizers==0.21.1
joblib==1.5.0
typing_extensions==4.13.2
typing-inspection==0.4.0
protobuf==6.31.0
pillow==11.2.1
networkx==3.4.2
cryptography==45.0.1
PyJWT==2.10.1
PyNaCl==1.5.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
